pyproject.toml
[project]
name = "multi-agent-ppm-platform"
version = "1.0.0"
description = "AI-native Multi-Agent Project Portfolio Management Platform"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {text = "MIT"}
authors = [
    {name = "Multi-Agent PPM Team"}
]
keywords = ["ppm", "project-management", "portfolio-management", "ai", "agents", "azure"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Runtime dependencies (monthly update cadence)
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "pydantic>=2.8.2,<3.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "python-multipart>=0.0.9,<0.1.0",

    # Azure SDKs
    "azure-identity>=1.17.1,<2.0.0",
    "azure-cosmos>=4.9.0,<5.0.0",
    "azure-storage-blob>=12.23.1,<13.0.0",
    "azure-servicebus>=7.13.2,<8.0.0",
    "azure-eventhub>=5.13.0,<6.0.0",
    "azure-keyvault-secrets>=4.8.0,<5.0.0",
    "azure-monitor-opentelemetry>=1.6.5,<2.0.0",
    "azure-monitor-query>=1.4.0,<2.0.0",
    "azure-ai-anomalydetector>=3.0.0,<4.0.0",
    "opentelemetry-sdk>=1.28.2,<2.0.0",
    "opentelemetry-exporter-otlp>=1.28.2,<2.0.0",
    "opentelemetry-exporter-azuremonitor>=1.0.0b35,<2.0.0",
    "cryptography>=43.0.3,<46.0.0",

    # Database
    "sqlalchemy>=2.0.36,<2.1.0",
    "alembic>=1.14.0,<2.0.0",
    "asyncpg>=0.29.0,<1.0.0; python_version < '3.13'",  # PostgreSQL async driver (py313 uses psycopg)
    "aiosqlite>=0.20.0,<1.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "psycopg[binary]>=3.2.3,<4.0.0; python_version >= '3.13'",
    "redis>=5.2.0,<6.0.0",

    # AI/ML
    "openai>=1.52.2,<2.0.0",
    "langchain>=0.3.9,<0.4.0",
    "tiktoken>=0.8.0,<1.0.0",
    "numpy>=1.26.4,<3.0.0",
    "faiss-cpu>=1.9.0,<2.0.0; platform_system != 'Windows'",

    # Utilities
    "python-dotenv>=1.0.1,<2.0.0",
    "python-dateutil>=2.9.0.post0,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "httpx>=0.27.2,<1.0.0",
    "aiofiles>=24.1.0,<25.0.0",
    "jsonschema>=4.23.0,<5.0.0",
    "pyjwt>=2.9.0,<3.0.0",
    "prometheus-client>=0.21.0,<1.0.0",
]

[project.optional-dependencies]
# Dev tooling dependencies (bi-weekly update cadence)
dev = [
    "black>=24.8.0,<25.0.0",
    "ruff>=0.7.0,<1.0.0",
    "mypy>=1.11.2,<2.0.0",
    "pre-commit>=4.0.1,<5.0.0",
    "bandit>=1.7.10,<2.0.0",
    "pip-audit>=2.7.3,<3.0.0",
]

# Test dependencies (bi-weekly update cadence)
test = [
    "pytest>=8.3.3,<9.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
    "pytest-timeout>=2.3.1,<3.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "httpx>=0.27.2,<1.0.0",  # For testing API
    "jsonschema>=4.23.0,<5.0.0",
]

docs = [
    "mkdocs>=1.6.1,<2.0.0",
    "mkdocs-material>=9.5.39,<10.0.0",
    "mkdocstrings[python]>=0.27.0,<1.0.0",
]

[tool.dependency_management]
runtime_update_cadence = "monthly"
dev_update_cadence = "bi-weekly"
test_update_cadence = "bi-weekly"
known_constraints = [
    "azure-ai-anomalydetector is held on 3.x because SDK v4 is not generally available.",
    "langchain remains constrained to 0.3.x to avoid breaking API changes expected in 1.x.",
    "sqlalchemy is capped at <2.1 pending async engine migration validation.",
]

[project.urls]
Homepage = "https://github.com/georgeprowse91/multi-agent-ppm-platform"
Documentation = "https://github.com/georgeprowse91/multi-agent-ppm-platform/tree/main/docs"
Repository = "https://github.com/georgeprowse91/multi-agent-ppm-platform"
Issues = "https://github.com/georgeprowse91/multi-agent-ppm-platform/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agents", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
timeout = 60
timeout_method = "thread"
addopts = "-v -ra --durations=10 --import-mode=importlib"
markers = [
    "asyncio: mark async tests handled by pytest-asyncio or fallback runner",
    "timeout(timeout): override the default timeout for a test or module",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "G"]
ignore = ["E501", "UP017"]  # Line too long (handled by black); UTC alias not in Py3.10

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports
"**/main.py" = ["E402"]  # Module imports after path manipulation
"**/legacy_main.py" = ["E402"]  # Legacy main files with path manipulation
"**/routes/*.py" = ["E402"]  # Route files with path bootstrapping
"tests/**/*.py" = ["E402", "N806"]  # Tests with dynamic imports and class loading
"**/tests/**/*.py" = ["E402", "N806"]  # Tests in subdirectories
"**/config.py" = ["E402"]  # Config files with path manipulation
"**/auth.py" = ["E402"]  # Auth files with path manipulation

[tool.mypy]
python_version = "3.11"
ignore_errors = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
exclude = [
  '(apps|services|agents)/.*/src/main\\.py',
  '(apps|services|agents)/.*/main\\.py',
  'services/agent-config/src/__init__\\.py',
  'packages/ui-kit/src/__init__\\.py',
  'services/data-lineage-service/src/retention_scheduler\\.py',
  'services/data-service/src/retention_scheduler\\.py',
  '(apps|services|agents)/.*/src/(storage|persistence|config|models|auth|metrics|circuit_breaker|leader_election|retention_scheduler|orchestrator|dependencies|connectors|analytics|documents|health|lineage|prompts|scope_research|assistant|audit|workflow|workspace|web_search|router)\\.py',
  '^(email_validator|events|prompt_registry|pydantic_settings|requests|runtime_flags)\\.py$',
  '^(sqlalchemy|integrations)/.*',
  '.*/tests/.*',
]

[[tool.mypy.overrides]]
module = ["fastapi.*", "pydantic.*", "uvicorn.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["yaml", "yaml.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["opentelemetry.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["numpy", "numpy.*"]
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["email_validator", "email_validator.*"]
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["faiss", "faiss.*"]
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["prometheus_client", "prometheus_client.*"]
follow_imports = "skip"
